context - track `InputFocusedContext` from `platform`
