+++ src/vs/workbench/browser/contextkeys.ts
@@ -3,12 +3,10 @@
  *  Licensed under the MIT License. See License.txt in the project root for license information.
  *--------------------------------------------------------------------------------------------*/
 
-import { Event } from '../../base/common/event.js';
-import { Disposable, DisposableStore, MutableDisposable } from '../../base/common/lifecycle.js';
+import { Disposable } from '../../base/common/lifecycle.js';
 import { IContextKeyService, IContextKey, setConstant as setConstantContextKey } from '../../platform/contextkey/common/contextkey.js';
-import { InputFocusedContext, IsMacContext, IsLinuxContext, IsWindowsContext, IsWebContext, IsMacNativeContext, IsDevelopmentContext, IsIOSContext, ProductQualityContext, IsMobileContext } from '../../platform/contextkey/common/contextkeys.js';
+import { IsMacContext, IsLinuxContext, IsWindowsContext, IsWebContext, IsMacNativeContext, IsDevelopmentContext, IsIOSContext, ProductQualityContext, IsMobileContext } from '../../platform/contextkey/common/contextkeys.js';
 import { SplitEditorsVertically, InEditorZenModeContext, AuxiliaryBarVisibleContext, SideBarVisibleContext, PanelAlignmentContext, PanelMaximizedContext, PanelVisibleContext, EmbedderIdentifierContext, EditorTabsVisibleContext, IsMainEditorCenteredLayoutContext, MainEditorAreaVisibleContext, DirtyWorkingCopiesContext, EmptyWorkspaceSupportContext, EnterMultiRootWorkspaceSupportContext, HasWebFileSystemAccess, IsMainWindowFullscreenContext, OpenFolderWorkspaceSupportContext, RemoteNameContext, VirtualWorkspaceContext, WorkbenchStateContext, WorkspaceFolderCountContext, PanelPositionContext, TemporaryWorkspaceContext, TitleBarVisibleContext, TitleBarStyleContext, IsAuxiliaryWindowFocusedContext, ActiveEditorGroupEmptyContext, ActiveEditorGroupIndexContext, ActiveEditorGroupLastContext, ActiveEditorGroupLockedContext, MultipleEditorGroupsContext, EditorsVisibleContext, AuxiliaryBarMaximizedContext, InAutomationContext } from '../common/contextkeys.js';
-import { trackFocus, addDisposableListener, EventType, onDidRegisterWindow, getActiveWindow, isEditableElement } from '../../base/browser/dom.js';
 import { preferredSideBySideGroupDirection, GroupDirection, IEditorGroupsService } from '../services/editor/common/editorGroupsService.js';
 import { IConfigurationService } from '../../platform/configuration/common/configuration.js';
 import { IWorkbenchEnvironmentService } from '../services/environment/common/environmentService.js';
@@ -27,7 +25,6 @@ import { isFullscreen, onDidChangeFullscreen } from '../../base/browser/browser.
 import { IEditorService } from '../services/editor/common/editorService.js';
 
 export class WorkbenchContextKeysHandler extends Disposable {
-	private inputFocusedContext: IContextKey<boolean>;
 
 	private dirtyWorkingCopiesContext: IContextKey<boolean>;
 
@@ -128,9 +125,6 @@ export class WorkbenchContextKeysHandler extends Disposable {
 		this.dirtyWorkingCopiesContext = DirtyWorkingCopiesContext.bindTo(this.contextKeyService);
 		this.dirtyWorkingCopiesContext.set(this.workingCopyService.hasDirty);
 
-		// Inputs
-		this.inputFocusedContext = InputFocusedContext.bindTo(this.contextKeyService);
-
 		// Workbench State
 		this.workbenchStateContext = WorkbenchStateContext.bindTo(this.contextKeyService);
 		this.updateWorkbenchStateContextKey();
@@ -224,14 +218,6 @@ export class WorkbenchContextKeysHandler extends Disposable {
 
 		this._register(this.editorGroupService.onDidChangeEditorPartOptions(() => this.updateEditorAreaContextKeys()));
 
-		this._register(Event.runAndSubscribe(onDidRegisterWindow, ({ window, disposables }) => {
-			const onFocusDisposables = disposables.add(new MutableDisposable<DisposableStore>());
-			disposables.add(addDisposableListener(window, EventType.FOCUS_IN, () => {
-				onFocusDisposables.value = new DisposableStore();
-				this.updateInputContextKeys(window.document, onFocusDisposables.value);
-			}, true));
-		}, { window: mainWindow, disposables: this._store }));
-
 		this._register(this.contextService.onDidChangeWorkbenchState(() => this.updateWorkbenchStateContextKey()));
 		this._register(this.contextService.onDidChangeWorkspaceFolders(() => {
 			this.updateWorkspaceFolderCountContextKey();
@@ -314,36 +300,6 @@ export class WorkbenchContextKeysHandler extends Disposable {
 
 	private updateEditorAreaContextKeys(): void {
 		this.editorTabsVisibleContext.set(this.editorGroupService.partOptions.showTabs === 'multiple');
-	}
-
-	private updateInputContextKeys(ownerDocument: Document, disposables: DisposableStore): void {
-
-		function activeElementIsInput(): boolean {
-			return !!ownerDocument.activeElement && isEditableElement(ownerDocument.activeElement);
-		}
-
-		const isInputFocused = activeElementIsInput();
-		this.inputFocusedContext.set(isInputFocused);
-
-		if (isInputFocused) {
-			const tracker = disposables.add(trackFocus(ownerDocument.activeElement as HTMLElement));
-			Event.once(tracker.onDidBlur)(() => {
-
-				// Ensure we are only updating the context key if we are
-				// still in the same document that we are tracking. This
-				// fixes a race condition in multi-window setups where
-				// the blur event arrives in the inactive window overwriting
-				// the context key of the active window. This is because
-				// blur events from the focus tracker are emitted with a
-				// timeout of 0.
-
-				if (getActiveWindow().document === ownerDocument) {
-					this.inputFocusedContext.set(activeElementIsInput());
-				}
-
-				tracker.dispose();
-			}, undefined, disposables);
-		}
 	}
 
 	private updateWorkbenchStateContextKey(): void {
